@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
+using System.Text.Json.Serialization;
 using Application.Comments;
-using Newtonsoft.Json;
 
 namespace Application.Activities
 {
@@ -15,7 +15,7 @@ namespace Application.Activities
         public string City { get; set; }
         public string Venue { get; set; }
 
-        [JsonProperty("attendees")]
+        [JsonPropertyName("attendees")]
         public ICollection<AttendeeDto> UserActivities { get; set; }
         public ICollection<CommentDto> Comments { get; set; }
     }
