@@ -51,7 +51,7 @@ namespace Application.Activities
                     Activity = activity,
                     AppUser = user,
                     IsHost = false,
-                    DateJoined = DateTime.Now
+                    DateJoined = DateTime.UtcNow
                 };
 
                 _context.UserActivities.Add(attendance);
