@@ -4,7 +4,6 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc.Filters;
 using Persistence;
 
 namespace Infrastructure.Security
@@ -25,21 +24,16 @@ namespace Infrastructure.Security
 
         protected override Task HandleRequirementAsync(AuthorizationHandlerContext context, IsHostRequirement requirement)
         {
-            if (context.Resource is AuthorizationFilterContext authContext)
-            {
-                var currentUserName = _httpContextAccessor.HttpContext.User?.Claims?.SingleOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
+            var currentUserName = _httpContextAccessor.HttpContext.User?.Claims?.SingleOrDefault(x => x.Type == ClaimTypes.NameIdentifier)?.Value;
 
-                var activityId = Guid.Parse(authContext.RouteData.Values["id"].ToString());
+            var activityId = Guid.Parse(_httpContextAccessor.HttpContext.Request.RouteValues.SingleOrDefault(x => x.Key == "id").Value.ToString());
 
-                var activity = _context.Activities.FindAsync(activityId).Result;
+            var activity = _context.Activities.FindAsync(activityId).Result;
 
-                var host = activity.UserActivities.FirstOrDefault(x => x.IsHost);
+            var host = activity.UserActivities.FirstOrDefault(x => x.IsHost);
 
-                if (host?.AppUser?.UserName == currentUserName)
-                    context.Succeed(requirement);
-            } else {
-                context.Fail();
-            }
+            if (host?.AppUser?.UserName == currentUserName)
+                context.Succeed(requirement);
 
             return Task.CompletedTask;
         }
