@@ -4,10 +4,11 @@ import { FormFieldProps, Form, Label } from 'semantic-ui-react';
 import {DateTimePicker} from 'react-widgets';
 
 interface IProps
-  extends FieldRenderProps<Date, HTMLInputElement>,
+  extends FieldRenderProps<Date, HTMLElement>,
     FormFieldProps {}
 
 const DateInput: React.FC<IProps> = ({
+    id = null,
     input,
     width,
     placeholder,
