@@ -13,7 +13,7 @@ const ErrorMessage: React.FC<IProps> = ({ error, text }) => {
       <Message.Header>{error.statusText}</Message.Header>
       {error.data && Object.keys(error.data.errors).length > 0 && (
         <Message.List>
-          {Object.values(error.data.errors).flat().map((err, i) => (
+          {Object.values(error.data.errors).flat().map((err: any, i) => (
             <Message.Item key={i}>{err}</Message.Item>
           ))}
         </Message.List>
