@@ -3,7 +3,7 @@ import { FieldRenderProps } from 'react-final-form';
 import { FormFieldProps, Form, Label } from 'semantic-ui-react';
 
 interface IProps
-  extends FieldRenderProps<string, HTMLTextAreaElement>,
+  extends FieldRenderProps<string, HTMLElement>,
     FormFieldProps {}
 
 const TextAreaInput: React.FC<IProps> = ({
