@@ -1,4 +1,4 @@
-import { observable, action, computed, runInAction, reaction } from 'mobx';
+import { observable, action, computed, runInAction, reaction, toJS } from 'mobx';
 import { SyntheticEvent } from 'react';
 import { IActivity } from '../models/activity';
 import agent from '../api/agent';
@@ -6,7 +6,7 @@ import { history } from '../..';
 import { toast } from 'react-toastify';
 import { RootStore } from './rootStore';
 import { setActivityProps, createAttendee } from '../common/util/util';
-import { HubConnection, HubConnectionBuilder, LogLevel } from '@aspnet/signalr';
+import {HubConnection, HubConnectionBuilder, LogLevel} from '@microsoft/signalr';
 
 const LIMIT = 2;
 
@@ -65,9 +65,9 @@ export default class ActivityStore {
     this.page = page;
   }
 
-  @action createHubConnection = () => {
+  @action createHubConnection = (activityId: string) => {
     this.hubConnection = new HubConnectionBuilder()
-      .withUrl('http://localhost:5000/chat', {
+      .withUrl(process.env.REACT_APP_API_CHAT_URL!, {
         accessTokenFactory: () => this.rootStore.commonStore.token!
       })
       .configureLogging(LogLevel.Information)
@@ -76,6 +76,10 @@ export default class ActivityStore {
     this.hubConnection
       .start()
       .then(() => console.log(this.hubConnection!.state))
+      .then(() => {
+        console.log('Attempting to join group');
+        this.hubConnection!.invoke('AddToGroup', activityId)
+      })
       .catch(error => console.log('Error establishing connection: ', error));
 
     this.hubConnection.on('ReceiveComment', comment => {
@@ -83,10 +87,19 @@ export default class ActivityStore {
         this.activity!.comments.push(comment)
       })
     })
+
+    this.hubConnection.on('Send', message => {
+      toast.info(message);
+    })
   };
 
   @action stopHubConnection = () => {
-    this.hubConnection!.stop()
+    this.hubConnection!.invoke('RemoveFromGroup', this.activity!.id)
+      .then(() => {
+        this.hubConnection!.stop()
+      })
+      .then(() => console.log('Connection stopped'))
+      .catch(err => console.log(err))
   }
 
   @action addComment = async (values: any) => {
@@ -97,6 +110,7 @@ export default class ActivityStore {
       console.log(error);
     }
   } 
+
 
   @computed get activitiesByDate() {
     return this.groupActivitiesByDate(
@@ -146,7 +160,7 @@ export default class ActivityStore {
     let activity = this.getActivity(id);
     if (activity) {
       this.activity = activity;
-      return activity;
+      return toJS(activity);
     } else {
       this.loadingInitial = true;
       try {
@@ -184,7 +198,6 @@ export default class ActivityStore {
       let attendees = [];
       attendees.push(attendee);
       activity.attendees = attendees;
-      activity.comments = [];
       activity.isHost = true;
       runInAction('create activity', () => {
         this.activityRegistry.set(activity.id, activity);
@@ -253,11 +266,11 @@ export default class ActivityStore {
           this.activityRegistry.set(this.activity.id, this.activity);
           this.loading = false;
         }
-      });
+      })
     } catch (error) {
       runInAction(() => {
         this.loading = false;
-      });
+      })
       toast.error('Problem signing up to activity');
     }
   };
@@ -275,11 +288,11 @@ export default class ActivityStore {
           this.activityRegistry.set(this.activity.id, this.activity);
           this.loading = false;
         }
-      });
+      })
     } catch (error) {
       runInAction(() => {
         this.loading = false;
-      });
+      })
       toast.error('Problem cancelling attendance');
     }
   };
