@@ -5,12 +5,14 @@ import { RootStore } from './rootStore';
 import { history } from '../..';
 
 export default class UserStore {
+  refreshTokenTimeout: any;
   rootStore: RootStore;
   constructor(rootStore: RootStore) {
     this.rootStore = rootStore;
   }
 
   @observable user: IUser | null = null;
+  @observable loading = false;
 
   @computed get isLoggedIn() {
     return !!this.user;
@@ -23,6 +25,7 @@ export default class UserStore {
         this.user = user;
       });
       this.rootStore.commonStore.setToken(user.token);
+      this.startRefreshTokenTimer(user);
       this.rootStore.modalStore.closeModal();
       history.push('/activities');
     } catch (error) {
@@ -32,12 +35,25 @@ export default class UserStore {
 
   @action register = async (values: IUserFormValues) => {
     try {
-      const user = await agent.User.register(values);
-      this.rootStore.commonStore.setToken(user.token);
+      await agent.User.register(values);
       this.rootStore.modalStore.closeModal();
-      history.push('/activities')
+      history.push(`/user/registerSuccess?email=${values.email}`)
     } catch (error) {
       throw error;
+    }
+  }
+
+  @action refreshToken = async () => {
+    this.stopRefreshTokenTimer();
+    try {
+      const user = await agent.User.refreshToken();
+      runInAction(() => {
+        this.user = user;
+      })
+      this.rootStore.commonStore.setToken(user.token);
+      this.startRefreshTokenTimer(user);
+    } catch (error) {
+      console.log(error);
     }
   }
 
@@ -47,6 +63,8 @@ export default class UserStore {
       runInAction(() => {
         this.user = user;
       });
+      this.rootStore.commonStore.setToken(user.token);
+      this.startRefreshTokenTimer(user);
     } catch (error) {
       console.log(error);
     }
@@ -57,4 +75,33 @@ export default class UserStore {
     this.user = null;
     history.push('/');
   };
+
+  @action fbLogin = async (response: any) => {
+    this.loading = true;
+    try {
+      const user = await agent.User.fbLogin(response.accessToken);
+      runInAction(() => {
+        this.user = user;
+        this.rootStore.commonStore.setToken(user.token);
+        this.startRefreshTokenTimer(user);
+        this.rootStore.modalStore.closeModal();
+        this.loading = false;
+      })
+      history.push('/activities');
+    } catch (error) {
+      this.loading = false;
+      throw error;
+    }
+  }
+
+  private startRefreshTokenTimer(user: IUser) {
+    const jwtToken = JSON.parse(atob(user.token.split('.')[1]));
+    const expires = new Date(jwtToken.exp * 1000);
+    const timeout = expires.getTime() - Date.now() - (60 * 1000);
+    this.refreshTokenTimeout = setTimeout(this.refreshToken, timeout);
+  }
+
+  private stopRefreshTokenTimer() {
+    clearTimeout(this.refreshTokenTimeout);
+  }
 }
