@@ -13,7 +13,7 @@ const ActivityList: React.FC = () => {
       {activitiesByDate.map(([group, activities]) => (
         <Fragment key={group}>
           <Label size='large' color='blue'>
-            {format(group, 'eeee do MMMM')}
+            {format(new Date(group), 'eeee do MMMM')}
           </Label>
           <Item.Group divided>
             {activities.map(activity => (
