@@ -17,11 +17,11 @@ const ActivityDetailedChat = () => {
   } = rootStore.activityStore;
 
   useEffect(() => {
-    createHubConnection();
+    createHubConnection(activity!.id);
     return () => {
       stopHubConnection();
     }
-  }, [createHubConnection, stopHubConnection])
+  }, [createHubConnection, stopHubConnection, activity])
 
   return (
     <Fragment>
